update repository design
