+++ app/Services/TaskService.php
@@ -2,18 +2,13 @@
 
 namespace App\Services;
 
-use App\Contracts\TaskRepository;
-use App\Repositories\Cache\TaskCacheRepository;
-use App\Repositories\HTTP\TaskHTTPRepository;
+use App\Repositories\TaskRepository;
 
 class TaskService extends \App\Providers\AppServiceProvider
 {
-    /**
-     * @var TaskRepository
-     */
     protected $taskRepository;
 
-    public function __construct(TaskHTTPRepository $taskRepository)
+    public function __construct(TaskRepository $taskRepository)
     {
         $this->taskRepository = $taskRepository;
     }
